Repackage REST client with Actor. Remove Runner in favor of Main for consistency
