+++ src/test/scala/edge/labs/orchestrator/jobs/rest/MockRestClient.scala
@@ -1,7 +1,5 @@
 package edge.labs.orchestrator.jobs.rest
 
-import edge.labs.orchestrator.rest.RestClient
-
 /* @author medge */
 
 /**
